refactor(gpu_gkr): make fused forward execution unconditional
