+++ gpu/gkr/Cargo.toml
@@ -37,4 +37,4 @@ worker = { workspace = true }
 fft = { workspace = true }
 rand = "0.9"
 itertools = "0.14"
-serde_json = "*"
+serde_json = "1"
