+++ gpu/gkr/native/gkr/eval_vm_isa.cuh
@@ -47,11 +47,6 @@ constexpr u32 FWD_VM_OPERAND_SMEM = 1;
 constexpr u32 FWD_VM_OPERAND_LDC = 2;
 constexpr u32 FWD_VM_OPERAND_SPECIAL = 3;
 constexpr u32 FWD_VM_SOURCE_WINDOW_SHIFT = FWD_VM_OPERAND_TAG_BITS;
-constexpr u32 FWD_VM_SOURCE_WINDOW_BITS = 6;
-constexpr u32 FWD_VM_SOURCE_WINDOW_MASK = (1u << FWD_VM_SOURCE_WINDOW_BITS) - 1;
-constexpr u32 FWD_VM_SOURCE_COLUMN_SHIFT = FWD_VM_SOURCE_WINDOW_SHIFT + FWD_VM_SOURCE_WINDOW_BITS;
-constexpr u32 FWD_VM_SOURCE_COLUMN_BITS = 7;
-constexpr u32 FWD_VM_SOURCE_COLUMN_MASK = (1u << FWD_VM_SOURCE_COLUMN_BITS) - 1;
 constexpr u32 FWD_VM_OPERAND_CELL_SHIFT = FWD_VM_OPERAND_TAG_BITS; // 2
 constexpr u32 FWD_VM_OPERAND_DESC_SHIFT = FWD_VM_OPERAND_TAG_BITS; // 2
 constexpr u32 FWD_VM_LDC_SUB_SHIFT = FWD_VM_OPERAND_TAG_BITS;      // 2
