+++ gpu/gkr/src/forward/vm/lower.rs
@@ -1,60 +1,38 @@
-//! Lowers a compiled forward layer to its CUDA descriptor.
-//!
-//! Compile-time columns are dense backing-table indices. Lowering resolves
-//! them to storage columns and splits each compile slot by distinct
-//! `(matrix_base, stride_bytes)` so every wire slot names one homogeneous
-//! matrix.
-use std::{
-    collections::{BTreeMap, BTreeSet},
-    ptr,
-};
+//! Lowers a forward circuit into its by-value CUDA descriptor.
 
-use gpu_gkr_compiler::{
-    encode_forward_program as encode, virtual_setup_kind_code, CompiledLayer,
-    ForwardDstLine as DstLine, ForwardEncodeError as EncodeError, ForwardInstr as Instr,
-    ForwardOperandField as OperandField, ForwardOperandLine as OperandLine,
-    ForwardProgram as Program, ForwardSpecialStrategy as SpecialStrategy,
-    FORWARD_MAX_COLS as MAX_COLS, FORWARD_SOURCE_WINDOW_COLUMNS as SOURCE_WINDOW_COLUMNS,
-};
+use std::{collections::BTreeMap, ptr};
 
-use crate::upstream::{ChallengeRef, GKRAddress, PrimeField, RangeWidth, ReadPlace};
 use gpu_core::primitives::field::{BF, E4};
+use gpu_gkr_compiler::{
+    encode_forward_runtime_program, virtual_setup_kind_code, CompiledLayer,
+    ForwardDstLine as DstLine, ForwardEncodeError as EncodeError, ForwardInstr as Instr,
+    ForwardLdcSub as LdcSub, ForwardOperandField as OperandField,
+    ForwardOperandLine as OperandLine, ForwardProgram as Program,
+    ForwardSpecialStrategy as SpecialStrategy, FORWARD_MAX_COLS as MAX_COLS,
+};
 
 use super::desc::{
-    pack_desc, FwdVmDesc, ARENA_GENERIC_FAMILY, ARENA_RANGE_CHECK_16, ARENA_TIMESTAMP,
-    ARG_DERIVED_E4_CAP, CONST_CAP, CONST_DERIVED_E4_CAP, DESC_CAP, DST_SLOT_COUNT,
+    pack_desc, FwdVmDesc, FwdVmLayer, ARENA_GENERIC_FAMILY, ARENA_RANGE_CHECK_16, ARENA_TIMESTAMP,
+    ARG_DERIVED_E4_CAP, CONST_CAP, CONST_DERIVED_E4_CAP, DESC_CAP, DST_SLOT_COUNT, LAYER_CAP,
     MAPPING_ARENA_COUNT, PROGRAM_CAP, SD_AGGREGATE, SD_DECODER, SD_INITS_TOP_BITS, SD_SETUP,
     SD_SINGLE_COLUMN, SD_VIRTUAL, SOURCE_WINDOW_COUNT,
 };
+use crate::upstream::{ChallengeRef, GKRAddress, PrimeField, RangeWidth, ReadPlace};
 
-/// One resolved storage column: the column's device pointer plus the geometry
-/// of the consolidated homogeneous matrix it belongs to (`storage/views.rs`:
-/// per-(layer, class, field) backing, column `poly_idx` at
-/// `matrix_base + poly_idx * stride_bytes`).
 #[derive(Clone, Copy, Debug)]
 pub(crate) struct ResolvedColumn {
-    /// Storage field of the matrix (16-B `E4` columns vs 4-B `BF` columns).
     pub is_e4: bool,
-    /// Device pointer of THIS column (`matrix_base + poly_idx * stride_bytes`).
     pub ptr: *const u8,
-    /// Device pointer of the consolidated matrix the column lives in.
     pub matrix_base: *mut u8,
-    /// Matrix column stride in bytes.
     pub stride_bytes: u32,
 }
 
-/// Per-layer inputs sourced from prover buffers. Unused pointers may be null.
-///
-/// The decoder fill value occupies the final const-derived-E4 bank slot.
 #[derive(Clone, Copy, Debug)]
-pub(crate) struct FwdVmHeaderInputs<'a> {
+pub(crate) struct FwdVmInputs<'a> {
     pub mapping_arena: [*const u32; MAPPING_ARENA_COUNT],
-    /// `generic_family` column of the decoder mapping (`num_generic_sets`,
-    /// the arena's last column). Required iff the layer has a `PeekDecoder`.
     pub decoder_mapping_col: Option<u16>,
     pub table: *const E4,
     pub table_len: u32,
-    /// Rows (= trace_len = mapping-arena column stride).
     pub count: u32,
     pub inits_and_teardowns_top_bits: &'a [u32],
 }
@@ -65,13 +43,11 @@ pub(crate) struct FwdVmHeaderInputs<'a> {
 )]
 #[derive(Debug)]
 pub(crate) enum FwdVmLowerError {
-    /// Wire-format encode failed (cap guard inside `gpu_gkr_compiler`).
     Encode(EncodeError),
-    /// Program exceeds the inline descriptor capacity.
     ProgramOverflow {
         lanes: usize,
     },
-    GroupSourceWindowOverflow {
+    SourceWindowOverflow {
         required: usize,
     },
     MissingLookupAdditiveSlot,
@@ -91,29 +67,20 @@ pub(crate) enum FwdVmLowerError {
         desc: usize,
         set_index: usize,
     },
-    /// A slot column's address did not resolve to resident storage.
     UnresolvedColumn {
         slot: u8,
         col: u16,
         addr: GKRAddress,
     },
-    /// Storage field disagrees with the slot's `BackingKey::field`.
     SlotFieldMismatch {
         slot: u8,
         col: u16,
         expect_e4: bool,
         got_e4: bool,
     },
-    /// Splitting compile slots into per-matrix wire slots needs more than
-    /// `DST_SLOT_COUNT` wire slots (SLOT_BITS=4 on the wire; `slot`/`col` locate
-    /// the column whose fresh `(base, stride)` group did not fit).
     WireSlotOverflow {
         slot: u8,
         col: u16,
-    },
-    SourceWindowOverflow {
-        window: u8,
-        column: u16,
     },
     UnmappedSource {
         window: u8,
@@ -129,27 +96,15 @@ pub(crate) enum FwdVmLowerError {
         window: u8,
         column: u16,
     },
-    SourceColRemapCollision {
-        window: u8,
-        matrix_col: usize,
-    },
-    /// Column pointer is not `matrix_base + k * stride_bytes` for integer `k`.
-    ColumnOffStride {
-        slot: u8,
-        col: u16,
-    },
-    /// Matrix column index exceeds the 10-bit wire `col` field.
     MatrixColOverflow {
         slot: u8,
         col: u16,
         matrix_col: usize,
     },
-    /// The program references a `(slot, col)` the backing table never assigned.
     UnmappedGlobal {
         slot: u8,
         col: u16,
     },
-    /// A special needs a mapping arena the header did not provide.
     MissingMappingArena {
         arena: u32,
     },
@@ -161,14 +116,7 @@ pub(crate) enum FwdVmLowerError {
     DecoderPredicateNotBase {
         addr: GKRAddress,
     },
-    /// Two decoder descs resolved to different execute-predicate columns
-    /// (the desc header holds ONE mask pointer).
     DecoderMaskConflict,
-    /// Distinct dense columns mapped to the same wire matrix column.
-    ColRemapCollision {
-        slot: u8,
-        matrix_col: usize,
-    },
 }
 
 pub(crate) fn read_place_to_gkr_address(place: &ReadPlace) -> GKRAddress {
@@ -182,314 +130,188 @@ pub(crate) fn read_place_to_gkr_address(place: &ReadPlace) -> GKRAddress {
     }
 }
 
+const SOURCE_COLUMN_BITS: u32 = 9;
+
+#[derive(Clone, Copy, Debug, Eq, Ord, PartialEq, PartialOrd)]
+struct SourceUseKey {
+    layer: usize,
+    window: u8,
+    column: u16,
+}
+
+#[derive(Clone, Copy, Debug)]
+struct PhysicalSourceUse {
+    key: SourceUseKey,
+    is_e4: bool,
+    matrix_base: *mut u8,
+    stride_bytes: u32,
+    matrix_column: usize,
+}
+
 struct SourceGeometry {
     base: [*mut u8; SOURCE_WINDOW_COUNT],
     stride_bytes: [u32; SOURCE_WINDOW_COUNT],
-    remap: BTreeMap<(u8, u16), (u8, u16)>,
+    remap: BTreeMap<SourceUseKey, (u8, u16)>,
     n_windows: usize,
 }
 
-fn source_coordinates(program: &Program) -> BTreeSet<(u8, u16)> {
-    let mut coordinates = BTreeSet::new();
-    let mut record = |operand: &OperandLine| {
-        if let OperandLine::Source { window, column, .. } = *operand {
-            coordinates.insert((window, column));
-        }
-    };
-    for instr in &program.instrs {
-        match instr {
-            Instr::Add { operands, .. } | Instr::Mul { operands, .. } => {
-                operands.iter().for_each(&mut record);
-            }
-            Instr::Fma { pairs, .. } => {
-                for (lhs, rhs) in pairs {
-                    record(lhs);
-                    record(rhs);
-                }
-            }
-            Instr::Mov { src: Some(src), .. } => record(src),
-            Instr::Mov { src: None, .. } => {}
-        }
-    }
-    coordinates
-}
+fn pack_source_geometry(uses: &[PhysicalSourceUse]) -> Result<SourceGeometry, FwdVmLowerError> {
+    let mut sorted = uses.to_vec();
+    sorted.sort_by_key(|source| {
+        (
+            source.is_e4,
+            source.matrix_base as usize,
+            source.stride_bytes,
+            source.matrix_column,
+        )
+    });
 
-fn derive_source_geometry(
-    cl: &CompiledLayer,
-    resolve_column: &dyn Fn(GKRAddress) -> Option<ResolvedColumn>,
-) -> Result<SourceGeometry, FwdVmLowerError> {
     let mut geometry = SourceGeometry {
-        base: [ptr::null_mut(); SOURCE_WINDOW_COUNT],
+        base: [std::ptr::null_mut(); SOURCE_WINDOW_COUNT],
         stride_bytes: [0; SOURCE_WINDOW_COUNT],
         remap: BTreeMap::new(),
         n_windows: 0,
     };
-    let coordinates = source_coordinates(&cl.program);
-    for compiler_window in 0..cl.source_windows.len() as u8 {
-        let expect_e4 = cl
-            .source_windows
-            .source_field(compiler_window)
-            .expect("dense source-window table")
-            == OperandField::Ext;
-        let mut groups = Vec::<(*mut u8, u32, Vec<(u16, usize)>)>::new();
-        for &(window, column) in coordinates
-            .iter()
-            .filter(|(window, _)| *window == compiler_window)
-        {
-            let place = cl
-                .source_windows
-                .resolve_read_place(window, column)
-                .ok_or(FwdVmLowerError::UnmappedSource { window, column })?;
-            let addr = read_place_to_gkr_address(&place);
-            let resolved = resolve_column(addr).ok_or(FwdVmLowerError::UnresolvedColumn {
-                slot: window,
-                col: column as u16,
-                addr,
-            })?;
-            if resolved.is_e4 != expect_e4 {
-                return Err(FwdVmLowerError::SourceFieldMismatch {
-                    window,
-                    column,
-                    expect_e4,
-                    got_e4: resolved.is_e4,
-                });
-            }
-            if resolved.stride_bytes == 0 {
-                return Err(FwdVmLowerError::SourceColumnOffStride { window, column });
-            }
-            let offset = (resolved.ptr as usize)
-                .checked_sub(resolved.matrix_base as usize)
-                .ok_or(FwdVmLowerError::SourceColumnOffStride { window, column })?;
-            if offset % resolved.stride_bytes as usize != 0 {
-                return Err(FwdVmLowerError::SourceColumnOffStride { window, column });
-            }
-            let matrix_col = offset / resolved.stride_bytes as usize;
-            let group = if let Some(index) = groups.iter().position(|(base, stride, _)| {
-                *base == resolved.matrix_base && *stride == resolved.stride_bytes
-            }) {
-                index
-            } else {
-                groups.push((resolved.matrix_base, resolved.stride_bytes, Vec::new()));
-                groups.len() - 1
-            };
-            if groups[group]
-                .2
-                .iter()
-                .any(|(_, existing)| *existing == matrix_col)
+    let mut active = None::<(bool, usize, u32, u8, usize)>;
+    for source in sorted {
+        let (window, first_column) = match active {
+            Some((is_e4, base, stride, window, first))
+                if is_e4 == source.is_e4
+                    && base == source.matrix_base as usize
+                    && stride == source.stride_bytes
+                    && source.matrix_column < first + (1 << SOURCE_COLUMN_BITS) =>
             {
-                return Err(FwdVmLowerError::SourceColRemapCollision { window, matrix_col });
+                (window, first)
             }
-            groups[group].2.push((column, matrix_col));
-        }
-
-        for (matrix_base, stride, mut columns) in groups {
-            columns.sort_by_key(|(_, matrix_col)| *matrix_col);
-            let mut active = None::<(u8, usize)>;
-            for (column, matrix_col) in columns {
-                let (wire, first_matrix_col) = match active {
-                    Some((wire, first)) if matrix_col < first + SOURCE_WINDOW_COLUMNS as usize => {
-                        (wire, first)
-                    }
-                    _ => {
-                        if geometry.n_windows >= SOURCE_WINDOW_COUNT {
-                            return Err(FwdVmLowerError::SourceWindowOverflow {
-                                window: compiler_window,
-                                column,
-                            });
-                        }
-                        let wire = geometry.n_windows as u8;
-                        let byte_offset = matrix_col
-                            .checked_mul(stride as usize)
-                            .and_then(|offset| (matrix_base as usize).checked_add(offset))
-                            .ok_or(FwdVmLowerError::SourceColumnOffStride {
-                                window: compiler_window,
-                                column,
-                            })?;
-                        geometry.base[geometry.n_windows] = byte_offset as *mut u8;
-                        geometry.stride_bytes[geometry.n_windows] = stride;
-                        geometry.n_windows += 1;
-                        active = Some((wire, matrix_col));
-                        (wire, matrix_col)
-                    }
-                };
-                geometry.remap.insert(
-                    (compiler_window, column),
-                    (wire, (matrix_col - first_matrix_col) as u16),
-                );
+            _ => {
+                if geometry.n_windows == SOURCE_WINDOW_COUNT {
+                    return Err(FwdVmLowerError::SourceWindowOverflow {
+                        required: geometry.n_windows + 1,
+                    });
+                }
+                let window = geometry.n_windows as u8;
+                let byte_offset = source
+                    .matrix_column
+                    .checked_mul(source.stride_bytes as usize)
+                    .and_then(|offset| (source.matrix_base as usize).checked_add(offset))
+                    .ok_or(FwdVmLowerError::SourceColumnOffStride {
+                        window: source.key.window,
+                        column: source.key.column,
+                    })?;
+                geometry.base[geometry.n_windows] = byte_offset as *mut u8;
+                geometry.stride_bytes[geometry.n_windows] = source.stride_bytes;
+                geometry.n_windows += 1;
+                active = Some((
+                    source.is_e4,
+                    source.matrix_base as usize,
+                    source.stride_bytes,
+                    window,
+                    source.matrix_column,
+                ));
+                (window, source.matrix_column)
             }
+        };
+        let coordinate = (
+            window,
+            u16::try_from(source.matrix_column - first_column)
+                .expect("source column fits nine bits"),
+        );
+        if let Some(previous) = geometry.remap.insert(source.key, coordinate) {
+            assert_eq!(previous, coordinate, "one logical source resolved two ways");
         }
     }
     Ok(geometry)
 }
 
-struct SlotGeometry {
-    base: [*mut u8; DST_SLOT_COUNT],
-    stride_bytes: [u32; DST_SLOT_COUNT],
-    n_wire_slots: usize,
-    remap: BTreeMap<(u8, u16), (u8, u16)>,
-    claimed: Vec<Vec<u16>>,
+#[derive(Clone, Copy, Debug, Eq, Ord, PartialEq, PartialOrd)]
+struct DestinationUseKey {
+    layer: usize,
+    slot: u8,
+    column: u16,
 }
 
-impl SlotGeometry {
-    /// Wire slots are not shared across compile slots.
-    fn wire_slot_for(
-        &mut self,
-        slot_groups: &mut Vec<(*mut u8, u32, u8)>,
-        resolved: &ResolvedColumn,
-        slot: u8,
-        col: u16,
-    ) -> Result<u8, FwdVmLowerError> {
-        if let Some(&(_, _, wire)) = slot_groups
-            .iter()
-            .find(|(b, s, _)| *b == resolved.matrix_base && *s == resolved.stride_bytes)
-        {
-            return Ok(wire);
-        }
-        if self.n_wire_slots >= DST_SLOT_COUNT {
-            return Err(FwdVmLowerError::WireSlotOverflow { slot, col });
-        }
-        let wire = self.n_wire_slots as u8;
-        self.base[self.n_wire_slots] = resolved.matrix_base;
-        self.stride_bytes[self.n_wire_slots] = resolved.stride_bytes;
-        self.n_wire_slots += 1;
-        slot_groups.push((resolved.matrix_base, resolved.stride_bytes, wire));
-        Ok(wire)
-    }
+#[derive(Clone, Copy)]
+struct LayerBankBases {
+    constant: u16,
+    arg_derived_e4: u16,
+    descriptor: u16,
+    lookup_additive: Option<u16>,
 }
 
-fn derive_slot_geometry(
-    cl: &CompiledLayer,
-    resolve_column: &dyn Fn(GKRAddress) -> Option<ResolvedColumn>,
-) -> Result<SlotGeometry, FwdVmLowerError> {
-    let backings = &cl.backings;
-    let mut geom = SlotGeometry {
-        base: [ptr::null_mut(); DST_SLOT_COUNT],
-        stride_bytes: [0; DST_SLOT_COUNT],
-        n_wire_slots: 0,
-        remap: BTreeMap::new(),
-        claimed: vec![Vec::new(); DST_SLOT_COUNT],
-    };
-    let dst_coordinates: BTreeSet<(u8, u16)> = cl
-        .program
-        .instrs
-        .iter()
-        .filter_map(|instr| match instr {
-            Instr::Mov {
-                dst: Some(DstLine::GlobalMaterialize { slot, col }),
-                ..
-            } => Some((*slot, *col)),
-            _ => None,
-        })
-        .collect();
-    for slot in 0..DST_SLOT_COUNT as u8 {
-        let slot_columns: Vec<u16> = dst_coordinates
-            .range((slot, 0)..=(slot, u16::MAX))
-            .map(|&(_, col)| col)
-            .collect();
-        if slot_columns.is_empty() {
-            continue;
-        }
-        let field = backings
-            .slot_field(slot)
-            .ok_or(FwdVmLowerError::UnmappedGlobal {
-                slot,
-                col: slot_columns[0],
-            })?;
-        let expect_e4 = field == OperandField::Ext;
-        // This compile slot's `(base, stride) -> wire slot` groups. Field
-        // homogeneity per wire slot follows from the per-column field check
-        // below: every group member matches THIS compile slot's field.
-        let mut slot_groups: Vec<(*mut u8, u32, u8)> = Vec::new();
-        for col in slot_columns {
-            let place = backings
-                .slot_col_to_read_place(slot, col)
-                .ok_or(FwdVmLowerError::UnmappedGlobal { slot, col })?;
-            let addr = read_place_to_gkr_address(&place);
-            let resolved = resolve_column(addr).ok_or(FwdVmLowerError::UnresolvedColumn {
-                slot,
-                col,
-                addr,
-            })?;
-            if resolved.is_e4 != expect_e4 {
-                return Err(FwdVmLowerError::SlotFieldMismatch {
-                    slot,
-                    col,
-                    expect_e4,
-                    got_e4: resolved.is_e4,
-                });
-            }
-            if resolved.stride_bytes == 0 {
-                return Err(FwdVmLowerError::ColumnOffStride { slot, col });
-            }
-            let wire = geom.wire_slot_for(&mut slot_groups, &resolved, slot, col)?;
-            let off = (resolved.ptr as usize)
-                .checked_sub(resolved.matrix_base as usize)
-                .ok_or(FwdVmLowerError::ColumnOffStride { slot, col })?;
-            if off % resolved.stride_bytes as usize != 0 {
-                return Err(FwdVmLowerError::ColumnOffStride { slot, col });
-            }
-            let matrix_col = off / resolved.stride_bytes as usize;
-            if matrix_col >= MAX_COLS as usize {
-                return Err(FwdVmLowerError::MatrixColOverflow {
-                    slot,
-                    col,
-                    matrix_col,
-                });
-            }
-            if geom.claimed[wire as usize].contains(&(matrix_col as u16)) {
-                return Err(FwdVmLowerError::ColRemapCollision { slot, matrix_col });
-            }
-            geom.claimed[wire as usize].push(matrix_col as u16);
-            geom.remap.insert((slot, col), (wire, matrix_col as u16));
-        }
-    }
-    Ok(geom)
-}
-
-fn remap_global(geom: &SlotGeometry, slot: u8, col: u16) -> Result<(u8, u16), FwdVmLowerError> {
-    geom.remap
-        .get(&(slot, col))
-        .copied()
-        .ok_or(FwdVmLowerError::UnmappedGlobal { slot, col })
-}
-
-fn remap_operand(geom: &SourceGeometry, o: OperandLine) -> Result<OperandLine, FwdVmLowerError> {
-    Ok(match o {
+fn rewrite_operand(
+    layer: usize,
+    operand: OperandLine,
+    sources: &BTreeMap<SourceUseKey, (u8, u16)>,
+    bases: LayerBankBases,
+) -> Result<OperandLine, FwdVmLowerError> {
+    Ok(match operand {
         OperandLine::Source { window, column } => {
-            let &(window, column) = geom
-                .remap
-                .get(&(window, column))
+            let &(window, column) = sources
+                .get(&SourceUseKey {
+                    layer,
+                    window,
+                    column,
+                })
                 .ok_or(FwdVmLowerError::UnmappedSource { window, column })?;
             OperandLine::Source { window, column }
         }
+        OperandLine::Ldc { sub, idx } => match sub {
+            LdcSub::Const => OperandLine::Ldc {
+                sub,
+                idx: bases.constant + idx,
+            },
+            LdcSub::ArgDerivedE4 => OperandLine::Ldc {
+                sub,
+                idx: bases.arg_derived_e4 + idx,
+            },
+            LdcSub::ConstDerivedE4 => {
+                assert_eq!(idx, 0, "unsupported const-derived-E4 index");
+                OperandLine::Ldc {
+                    sub,
+                    idx: bases
+                        .lookup_additive
+                        .ok_or(FwdVmLowerError::MissingLookupAdditiveSlot)?,
+                }
+            }
+            LdcSub::Special => operand,
+        },
+        OperandLine::Special { desc } => OperandLine::Special {
+            desc: bases.descriptor + desc,
+        },
         other => other,
     })
 }
 
-fn remap_dst(geom: &SlotGeometry, d: DstLine) -> Result<DstLine, FwdVmLowerError> {
-    Ok(match d {
+fn rewrite_dst(
+    layer: usize,
+    dst: DstLine,
+    destinations: &BTreeMap<DestinationUseKey, (u8, u16)>,
+) -> Result<DstLine, FwdVmLowerError> {
+    Ok(match dst {
         DstLine::GlobalMaterialize { slot, col } => {
-            let (slot, col) = remap_global(geom, slot, col)?;
+            let &(slot, col) = destinations
+                .get(&DestinationUseKey {
+                    layer,
+                    slot,
+                    column: col,
+                })
+                .ok_or(FwdVmLowerError::UnmappedGlobal { slot, col })?;
             DstLine::GlobalMaterialize { slot, col }
         }
         other => other,
     })
 }
 
-/// Rewrite every `Global` operand/dst from the backing table's dense
-/// `(compile slot, col)` to the storage `(wire slot, matrix col)` (module
-/// doc). Structure-preserving: only `Global`/`GlobalMaterialize` slot/col
-/// fields change, and the rewritten program goes back through the
-/// cap-guarded `encode` (`SlotOutOfRange` at `MAX_SLOTS`, `ColOutOfRange`
-/// at `MAX_COLS`), so a malformed remap cannot reach the wire.
 fn rewrite_program(
-    cl: &CompiledLayer,
-    source_geom: &SourceGeometry,
-    dst_geom: &SlotGeometry,
+    layer: usize,
+    program: &Program,
+    sources: &BTreeMap<SourceUseKey, (u8, u16)>,
+    destinations: &BTreeMap<DestinationUseKey, (u8, u16)>,
+    bases: LayerBankBases,
 ) -> Result<Program, FwdVmLowerError> {
-    let mut instrs = Vec::with_capacity(cl.program.instrs.len());
-    for instr in &cl.program.instrs {
-        instrs.push(match instr {
+    let mut instrs = Vec::with_capacity(program.instrs.len());
+    for instruction in &program.instrs {
+        instrs.push(match instruction {
             Instr::Add {
                 field,
                 sign,
@@ -499,7 +321,7 @@ fn rewrite_program(
                 sign: *sign,
                 operands: operands
                     .iter()
-                    .map(|o| remap_operand(source_geom, *o))
+                    .map(|operand| rewrite_operand(layer, *operand, sources, bases))
                     .collect::<Result<_, _>>()?,
             },
             Instr::Mul {
@@ -511,7 +333,7 @@ fn rewrite_program(
                 negate_acc: *negate_acc,
                 operands: operands
                     .iter()
-                    .map(|o| remap_operand(source_geom, *o))
+                    .map(|operand| rewrite_operand(layer, *operand, sources, bases))
                     .collect::<Result<_, _>>()?,
             },
             Instr::Fma {
@@ -525,10 +347,10 @@ fn rewrite_program(
                 sign: *sign,
                 pairs: pairs
                     .iter()
-                    .map(|(l, r)| {
+                    .map(|(lhs, rhs)| {
                         Ok((
-                            remap_operand(source_geom, *l)?,
-                            remap_operand(source_geom, *r)?,
+                            rewrite_operand(layer, *lhs, sources, bases)?,
+                            rewrite_operand(layer, *rhs, sources, bases)?,
                         ))
                     })
                     .collect::<Result<_, FwdVmLowerError>>()?,
@@ -541,158 +363,482 @@ fn rewrite_program(
             } => Instr::Mov {
                 dir: *dir,
                 field: *field,
-                dst: dst.map(|d| remap_dst(dst_geom, d)).transpose()?,
-                src: src.map(|s| remap_operand(source_geom, s)).transpose()?,
+                dst: dst
+                    .map(|dst| rewrite_dst(layer, dst, destinations))
+                    .transpose()?,
+                src: src
+                    .map(|src| rewrite_operand(layer, src, sources, bases))
+                    .transpose()?,
             },
         });
     }
     Ok(Program { instrs })
 }
 
-/// Assemble the by-value forward VM descriptor for one compiled layer.
-///
-/// - `resolve_column` maps a flat `GKRAddress` to its resident storage column
-///   (production: the consolidated `storage/views.rs` matrices; tests: mocks).
-/// - `challenge` resolves schedule-time `ArgDerivedE4` references.
-/// - `ConstDerivedE4` values and the optional decoder fill are staged by the
-///   production binding before the layer launch; only their bank indices and
-///   count live in the descriptor.
-pub(crate) fn lower_layer_desc(
-    cl: &CompiledLayer,
-    header: &FwdVmHeaderInputs<'_>,
-    resolve_column: &dyn Fn(GKRAddress) -> Option<ResolvedColumn>,
-    challenge: &dyn Fn(&ChallengeRef) -> E4,
-) -> Result<FwdVmDesc, FwdVmLowerError> {
-    // SAFETY: all-zero bytes are a valid `FwdVmDesc` — plain-old-data fields
-    // plus nullable raw pointers; every meaningful field is filled below.
-    let mut desc: FwdVmDesc = unsafe { core::mem::zeroed() };
+fn empty_desc() -> FwdVmDesc {
+    // SAFETY: the descriptor is plain data and all pointer fields may be null.
+    unsafe { core::mem::zeroed() }
+}
 
-    // ----- column geometry + program rewrite + encode. -----
-    let source_geom = derive_source_geometry(cl, resolve_column)?;
-    let dst_geom = derive_slot_geometry(cl, resolve_column)?;
-    let program = rewrite_program(cl, &source_geom, &dst_geom)?;
-    let lanes = encode(&program).map_err(FwdVmLowerError::Encode)?;
-    desc.source_base = source_geom.base;
-    desc.source_stride_bytes = source_geom.stride_bytes;
-    desc.dst_base = dst_geom.base;
-    desc.dst_stride_bytes = dst_geom.stride_bytes;
-    desc.n_instr = program.instrs.len() as u32;
+fn append_program(
+    desc: &mut FwdVmDesc,
+    next_lane: &mut usize,
+    layer: usize,
+    program: &Program,
+) -> Result<(), FwdVmLowerError> {
+    assert!(layer < LAYER_CAP);
+    let lanes = encode_forward_runtime_program(program).map_err(FwdVmLowerError::Encode)?;
+    let end = next_lane
+        .checked_add(lanes.len())
+        .ok_or(FwdVmLowerError::ProgramOverflow { lanes: usize::MAX })?;
+    if end > PROGRAM_CAP {
+        return Err(FwdVmLowerError::ProgramOverflow { lanes: end });
+    }
+    desc.layers[layer] = FwdVmLayer {
+        program_offset: u16::try_from(*next_lane).expect("PROGRAM_CAP fits u16"),
+        instruction_count: u16::try_from(program.instrs.len())
+            .expect("forward instruction count fits u16"),
+    };
+    desc.program[*next_lane..end].copy_from_slice(&lanes);
+    *next_lane = end;
+    Ok(())
+}
 
-    if lanes.len() > PROGRAM_CAP {
-        return Err(FwdVmLowerError::ProgramOverflow { lanes: lanes.len() });
-    }
-    desc.program[..lanes.len()].copy_from_slice(&lanes);
+#[derive(Clone, Copy, Debug)]
+struct PhysicalDestinationUse {
+    key: DestinationUseKey,
+    is_e4: bool,
+    matrix_base: *mut u8,
+    stride_bytes: u32,
+    matrix_column: usize,
+}
 
-    // ----- banks (hard caps, no fallback). -----
-    let consts = cl.consts.values();
-    if consts.len() > CONST_CAP {
-        return Err(FwdVmLowerError::ConstBankOverflow { n: consts.len() });
-    }
-    for (i, &v) in consts.iter().enumerate() {
-        desc.consts[i] = BF::from_u32_with_reduction(v);
-    }
-    let mut n_consts = consts.len();
+struct DestinationGeometry {
+    base: [*mut u8; DST_SLOT_COUNT],
+    stride_bytes: [u32; DST_SLOT_COUNT],
+    remap: BTreeMap<DestinationUseKey, (u8, u16)>,
+}
 
-    let n_arg = cl.derived_e4.arg_refs().len();
-    if n_arg > ARG_DERIVED_E4_CAP {
-        return Err(FwdVmLowerError::ArgDerivedE4Overflow { n: n_arg });
-    }
-    for (i, r) in cl.derived_e4.arg_refs().iter().enumerate() {
-        desc.arg_derived_e4[i] = challenge(r);
-    }
-    let n_const_derived_e4 = usize::from(cl.derived_e4.uses_lookup_additive())
-        + usize::from(
-            cl.specials
-                .iter()
-                .any(|special| matches!(special, SpecialStrategy::PeekDecoder { .. })),
-        );
-    if n_const_derived_e4 > CONST_DERIVED_E4_CAP {
-        return Err(FwdVmLowerError::ConstDerivedE4Overflow {
-            n: n_const_derived_e4,
-        });
-    }
-
-    // ----- special descriptors (packed u32 each) + header pointers. -----
-    let n_descs = cl.specials.len();
-    if n_descs > DESC_CAP {
-        return Err(FwdVmLowerError::DescOverflow { n: n_descs });
-    }
-    let mut uses_table = false;
-    let mut mask: *const BF = ptr::null();
-    let require_arena = |arena: u32| -> Result<u32, FwdVmLowerError> {
-        if header.mapping_arena[arena as usize].is_null() {
-            Err(FwdVmLowerError::MissingMappingArena { arena })
-        } else {
-            Ok(arena)
+fn pack_destination_geometry(
+    uses: &[PhysicalDestinationUse],
+) -> Result<DestinationGeometry, FwdVmLowerError> {
+    let mut sorted = uses.to_vec();
+    sorted.sort_by_key(|destination| {
+        (
+            destination.is_e4,
+            destination.matrix_base as usize,
+            destination.stride_bytes,
+            destination.matrix_column,
+        )
+    });
+    let mut geometry = DestinationGeometry {
+        base: [ptr::null_mut(); DST_SLOT_COUNT],
+        stride_bytes: [0; DST_SLOT_COUNT],
+        remap: BTreeMap::new(),
+    };
+    let mut active = None::<(bool, usize, u32, u8)>;
+    let mut n_slots = 0usize;
+    for destination in sorted {
+        if destination.matrix_column >= MAX_COLS as usize {
+            return Err(FwdVmLowerError::MatrixColOverflow {
+                slot: destination.key.slot,
+                col: destination.key.column,
+                matrix_col: destination.matrix_column,
+            });
         }
-    };
-    let set_index_u16 = |d: usize, set_index: usize| -> Result<u16, FwdVmLowerError> {
-        u16::try_from(set_index)
-            .map_err(|_| FwdVmLowerError::SetIndexOverflow { desc: d, set_index })
-    };
-    for (d, sd) in cl.specials.iter().enumerate() {
-        desc.descs[d] = match sd {
-            SpecialStrategy::PeekSingleColumn { set_index, width } => {
-                let arena = require_arena(match width {
-                    RangeWidth::Bits16 => ARENA_RANGE_CHECK_16,
-                    RangeWidth::Timestamp => ARENA_TIMESTAMP,
-                })?;
-                pack_desc(SD_SINGLE_COLUMN, arena, set_index_u16(d, *set_index)?, 0)
+        let slot = match active {
+            Some((is_e4, base, stride, slot))
+                if is_e4 == destination.is_e4
+                    && base == destination.matrix_base as usize
+                    && stride == destination.stride_bytes =>
+            {
+                slot
             }
-            SpecialStrategy::PeekAggregate { set_index } => {
-                uses_table = true;
-                let arena = require_arena(ARENA_GENERIC_FAMILY)?;
-                pack_desc(SD_AGGREGATE, arena, set_index_u16(d, *set_index)?, 0)
-            }
-            SpecialStrategy::PeekSetup => {
-                uses_table = true;
-                pack_desc(SD_SETUP, 0, 0, 0)
-            }
-            SpecialStrategy::PeekDecoder { predicate, .. } => {
-                uses_table = true;
-                let arena = require_arena(ARENA_GENERIC_FAMILY)?;
-                let col = header
-                    .decoder_mapping_col
-                    .ok_or(FwdVmLowerError::MissingDecoderMappingCol)?;
-                let addr = read_place_to_gkr_address(predicate);
-                let pred = resolve_column(addr)
-                    .ok_or(FwdVmLowerError::DecoderPredicateUnresolved { addr })?;
-                if pred.is_e4 {
-                    return Err(FwdVmLowerError::DecoderPredicateNotBase { addr });
+            _ => {
+                if n_slots == DST_SLOT_COUNT {
+                    return Err(FwdVmLowerError::WireSlotOverflow {
+                        slot: destination.key.slot,
+                        col: destination.key.column,
+                    });
                 }
-                let pred_ptr = pred.ptr as *const BF;
-                if mask.is_null() {
-                    mask = pred_ptr;
-                } else if mask != pred_ptr {
-                    return Err(FwdVmLowerError::DecoderMaskConflict);
-                }
-                pack_desc(SD_DECODER, arena, col, (CONST_DERIVED_E4_CAP - 1) as u32)
-            }
-            SpecialStrategy::VirtualSetup { kind } => {
-                // vkind = the NATIVE `gkr_base_source_kind` value VERBATIM:
-                // `KIND_ORDER` code + 2 (pinned by desc.rs const asserts).
-                pack_desc(SD_VIRTUAL, 0, 0, virtual_setup_kind_code(kind) + 2)
-            }
-            SpecialStrategy::InitsAndTeardownsTopBits { reference } => {
-                let raw = *header
-                    .inits_and_teardowns_top_bits
-                    .get(reference.set_index)
-                    .ok_or(FwdVmLowerError::SetIndexOverflow {
-                        desc: d,
-                        set_index: reference.set_index,
-                    })?;
-                if n_consts >= CONST_CAP {
-                    return Err(FwdVmLowerError::ConstBankOverflow { n: n_consts + 1 });
-                }
-                let raw = raw.checked_shl(reference.shift).unwrap_or(0);
-                desc.consts[n_consts] = BF::from_u32_with_reduction(raw);
-                let slot = u16::try_from(n_consts).expect("CONST_CAP fits u16");
-                n_consts += 1;
-                pack_desc(SD_INITS_TOP_BITS, 0, slot, 0)
+                let slot = n_slots as u8;
+                geometry.base[n_slots] = destination.matrix_base;
+                geometry.stride_bytes[n_slots] = destination.stride_bytes;
+                n_slots += 1;
+                active = Some((
+                    destination.is_e4,
+                    destination.matrix_base as usize,
+                    destination.stride_bytes,
+                    slot,
+                ));
+                slot
             }
         };
+        let coordinate = (slot, destination.matrix_column as u16);
+        if let Some(previous) = geometry.remap.insert(destination.key, coordinate) {
+            assert_eq!(
+                previous, coordinate,
+                "one logical destination resolved two ways"
+            );
+        }
     }
+    Ok(geometry)
+}
+
+fn physical_matrix_column(
+    resolved: ResolvedColumn,
+    slot: u8,
+    column: u16,
+) -> Result<usize, FwdVmLowerError> {
+    if resolved.stride_bytes == 0 {
+        return Err(FwdVmLowerError::SourceColumnOffStride {
+            window: slot,
+            column,
+        });
+    }
+    let offset = (resolved.ptr as usize)
+        .checked_sub(resolved.matrix_base as usize)
+        .ok_or(FwdVmLowerError::SourceColumnOffStride {
+            window: slot,
+            column,
+        })?;
+    if offset % resolved.stride_bytes as usize != 0 {
+        return Err(FwdVmLowerError::SourceColumnOffStride {
+            window: slot,
+            column,
+        });
+    }
+    Ok(offset / resolved.stride_bytes as usize)
+}
+
+fn visit_layer_source_operands(
+    layer_index: usize,
+    layer: &CompiledLayer,
+    mut visit: impl FnMut(SourceUseKey, OperandField) -> Result<(), FwdVmLowerError>,
+) -> Result<(), FwdVmLowerError> {
+    let mut record = |operand: &OperandLine, field: OperandField| {
+        if let OperandLine::Source { window, column } = *operand {
+            visit(
+                SourceUseKey {
+                    layer: layer_index,
+                    window,
+                    column,
+                },
+                field,
+            )?;
+        }
+        Ok(())
+    };
+    for instruction in &layer.program.instrs {
+        match instruction {
+            Instr::Add {
+                field, operands, ..
+            }
+            | Instr::Mul {
+                field, operands, ..
+            } => {
+                for operand in operands {
+                    record(operand, *field)?;
+                }
+            }
+            Instr::Fma {
+                field_lhs,
+                field_rhs,
+                pairs,
+                ..
+            } => {
+                for (lhs, rhs) in pairs {
+                    record(lhs, *field_lhs)?;
+                    record(rhs, *field_rhs)?;
+                }
+            }
+            Instr::Mov {
+                field,
+                src: Some(src),
+                ..
+            } => record(src, *field)?,
+            Instr::Mov { src: None, .. } => {}
+        }
+    }
+    Ok(())
+}
+
+pub(crate) struct LoweredFwdVm {
+    pub desc: FwdVmDesc,
+    pub lookup_additive_slot: Option<usize>,
+    pub decoder_fill_slot: Option<usize>,
+}
+
+pub(crate) fn lower_desc(
+    layers: &[CompiledLayer],
+    header: &FwdVmInputs<'_>,
+    resolve_column: &dyn Fn(GKRAddress) -> Option<ResolvedColumn>,
+    challenge: &dyn Fn(&ChallengeRef) -> E4,
+) -> Result<LoweredFwdVm, FwdVmLowerError> {
+    assert!(
+        !layers.is_empty(),
+        "forward VM must have at least one layer"
+    );
+    assert!(layers.len() <= LAYER_CAP, "forward VM exceeds LAYER_CAP");
+
+    let mut physical_sources = Vec::new();
+    let mut physical_destinations = Vec::new();
+    for (layer_index, layer) in layers.iter().enumerate() {
+        visit_layer_source_operands(layer_index, layer, |key, field| {
+            let source_field = layer
+                .source_windows
+                .source_field(key.window)
+                .expect("compiled source window has a field");
+            if source_field != field {
+                return Err(FwdVmLowerError::SourceFieldMismatch {
+                    window: key.window,
+                    column: key.column,
+                    expect_e4: field == OperandField::Ext,
+                    got_e4: source_field == OperandField::Ext,
+                });
+            }
+            let place = layer
+                .source_windows
+                .resolve_read_place(key.window, key.column)
+                .ok_or(FwdVmLowerError::UnmappedSource {
+                    window: key.window,
+                    column: key.column,
+                })?;
+            let address = read_place_to_gkr_address(&place);
+            let resolved = resolve_column(address).ok_or(FwdVmLowerError::UnresolvedColumn {
+                slot: key.window,
+                col: key.column,
+                addr: address,
+            })?;
+            let expect_e4 = field == OperandField::Ext;
+            if resolved.is_e4 != expect_e4 {
+                return Err(FwdVmLowerError::SourceFieldMismatch {
+                    window: key.window,
+                    column: key.column,
+                    expect_e4,
+                    got_e4: resolved.is_e4,
+                });
+            }
+            physical_sources.push(PhysicalSourceUse {
+                key,
+                is_e4: resolved.is_e4,
+                matrix_base: resolved.matrix_base,
+                stride_bytes: resolved.stride_bytes,
+                matrix_column: physical_matrix_column(resolved, key.window, key.column)?,
+            });
+            Ok(())
+        })?;
+
+        for instruction in &layer.program.instrs {
+            let Instr::Mov {
+                field,
+                dst: Some(DstLine::GlobalMaterialize { slot, col }),
+                ..
+            } = instruction
+            else {
+                continue;
+            };
+            let place = layer.backings.slot_col_to_read_place(*slot, *col).ok_or(
+                FwdVmLowerError::UnmappedGlobal {
+                    slot: *slot,
+                    col: *col,
+                },
+            )?;
+            let address = read_place_to_gkr_address(&place);
+            let resolved = resolve_column(address).ok_or(FwdVmLowerError::UnresolvedColumn {
+                slot: *slot,
+                col: *col,
+                addr: address,
+            })?;
+            let expect_e4 = *field == OperandField::Ext;
+            if resolved.is_e4 != expect_e4 {
+                return Err(FwdVmLowerError::SlotFieldMismatch {
+                    slot: *slot,
+                    col: *col,
+                    expect_e4,
+                    got_e4: resolved.is_e4,
+                });
+            }
+            physical_destinations.push(PhysicalDestinationUse {
+                key: DestinationUseKey {
+                    layer: layer_index,
+                    slot: *slot,
+                    column: *col,
+                },
+                is_e4: resolved.is_e4,
+                matrix_base: resolved.matrix_base,
+                stride_bytes: resolved.stride_bytes,
+                matrix_column: physical_matrix_column(resolved, *slot, *col)?,
+            });
+        }
+    }
+
+    let source_geometry = pack_source_geometry(&physical_sources)?;
+    let destination_geometry = pack_destination_geometry(&physical_destinations)?;
+    let uses_lookup_additive = layers
+        .iter()
+        .any(|layer| layer.derived_e4.uses_lookup_additive());
+    let uses_decoder_fill = layers.iter().any(|layer| {
+        layer
+            .specials
+            .iter()
+            .any(|special| matches!(special, SpecialStrategy::PeekDecoder { .. }))
+    });
+    let lookup_additive_slot = uses_lookup_additive.then_some(0usize);
+    let decoder_fill_slot = uses_decoder_fill.then_some(usize::from(uses_lookup_additive));
+    let n_const_derived = usize::from(uses_lookup_additive) + usize::from(uses_decoder_fill);
+    if n_const_derived > CONST_DERIVED_E4_CAP {
+        return Err(FwdVmLowerError::ConstDerivedE4Overflow { n: n_const_derived });
+    }
+
+    let mut desc = empty_desc();
+    desc.source_base = source_geometry.base;
+    desc.source_stride_bytes = source_geometry.stride_bytes;
+    desc.dst_base = destination_geometry.base;
+    desc.dst_stride_bytes = destination_geometry.stride_bytes;
     desc.mapping_arena = header.mapping_arena;
+    desc.count = header.count;
+    desc.layer_count = layers.len() as u32;
+
+    let mut n_consts = 0usize;
+    let mut n_args = 0usize;
+    let mut n_descs = 0usize;
+    let mut next_lane = 0usize;
+    let mut uses_table = false;
+    let mut mask: *const BF = ptr::null();
+    for (layer_index, layer) in layers.iter().enumerate() {
+        let constant_base = n_consts;
+        for &value in layer.consts.values() {
+            if n_consts == CONST_CAP {
+                return Err(FwdVmLowerError::ConstBankOverflow { n: n_consts + 1 });
+            }
+            desc.consts[n_consts] = BF::from_u32_with_reduction(value);
+            n_consts += 1;
+        }
+
+        let arg_base = n_args;
+        for reference in layer.derived_e4.arg_refs() {
+            if n_args == ARG_DERIVED_E4_CAP {
+                return Err(FwdVmLowerError::ArgDerivedE4Overflow { n: n_args + 1 });
+            }
+            desc.arg_derived_e4[n_args] = challenge(reference);
+            n_args += 1;
+        }
+
+        let descriptor_base = n_descs;
+        let require_arena = |arena: u32| -> Result<u32, FwdVmLowerError> {
+            if header.mapping_arena[arena as usize].is_null() {
+                Err(FwdVmLowerError::MissingMappingArena { arena })
+            } else {
+                Ok(arena)
+            }
+        };
+        for special in layer.specials.iter() {
+            if n_descs == DESC_CAP {
+                return Err(FwdVmLowerError::DescOverflow { n: n_descs + 1 });
+            }
+            let local_desc = n_descs - descriptor_base;
+            desc.descs[n_descs] = match special {
+                SpecialStrategy::PeekSingleColumn { set_index, width } => {
+                    let arena = require_arena(match width {
+                        RangeWidth::Bits16 => ARENA_RANGE_CHECK_16,
+                        RangeWidth::Timestamp => ARENA_TIMESTAMP,
+                    })?;
+                    pack_desc(
+                        SD_SINGLE_COLUMN,
+                        arena,
+                        u16::try_from(*set_index).map_err(|_| {
+                            FwdVmLowerError::SetIndexOverflow {
+                                desc: n_descs,
+                                set_index: *set_index,
+                            }
+                        })?,
+                        0,
+                    )
+                }
+                SpecialStrategy::PeekAggregate { set_index } => {
+                    uses_table = true;
+                    let arena = require_arena(ARENA_GENERIC_FAMILY)?;
+                    pack_desc(
+                        SD_AGGREGATE,
+                        arena,
+                        u16::try_from(*set_index).map_err(|_| {
+                            FwdVmLowerError::SetIndexOverflow {
+                                desc: n_descs,
+                                set_index: *set_index,
+                            }
+                        })?,
+                        0,
+                    )
+                }
+                SpecialStrategy::PeekSetup => {
+                    uses_table = true;
+                    pack_desc(SD_SETUP, 0, 0, 0)
+                }
+                SpecialStrategy::PeekDecoder { predicate } => {
+                    uses_table = true;
+                    let arena = require_arena(ARENA_GENERIC_FAMILY)?;
+                    let column = header
+                        .decoder_mapping_col
+                        .ok_or(FwdVmLowerError::MissingDecoderMappingCol)?;
+                    let address = read_place_to_gkr_address(predicate);
+                    let resolved = resolve_column(address)
+                        .ok_or(FwdVmLowerError::DecoderPredicateUnresolved { addr: address })?;
+                    if resolved.is_e4 {
+                        return Err(FwdVmLowerError::DecoderPredicateNotBase { addr: address });
+                    }
+                    let predicate_mask = resolved.ptr as *const BF;
+                    if mask.is_null() {
+                        mask = predicate_mask;
+                    } else if mask != predicate_mask {
+                        return Err(FwdVmLowerError::DecoderMaskConflict);
+                    }
+                    pack_desc(
+                        SD_DECODER,
+                        arena,
+                        column,
+                        decoder_fill_slot.expect("decoder slot exists") as u32,
+                    )
+                }
+                SpecialStrategy::VirtualSetup { kind } => {
+                    pack_desc(SD_VIRTUAL, 0, 0, virtual_setup_kind_code(kind) + 2)
+                }
+                SpecialStrategy::InitsAndTeardownsTopBits { reference } => {
+                    if n_consts == CONST_CAP {
+                        return Err(FwdVmLowerError::ConstBankOverflow { n: n_consts + 1 });
+                    }
+                    let raw = *header
+                        .inits_and_teardowns_top_bits
+                        .get(reference.set_index)
+                        .ok_or(FwdVmLowerError::SetIndexOverflow {
+                            desc: local_desc,
+                            set_index: reference.set_index,
+                        })?;
+                    desc.consts[n_consts] =
+                        BF::from_u32_with_reduction(raw.checked_shl(reference.shift).unwrap_or(0));
+                    let slot = n_consts as u16;
+                    n_consts += 1;
+                    pack_desc(SD_INITS_TOP_BITS, 0, slot, 0)
+                }
+            };
+            n_descs += 1;
+        }
+
+        let rewritten = rewrite_program(
+            layer_index,
+            &layer.program,
+            &source_geometry.remap,
+            &destination_geometry.remap,
+            LayerBankBases {
+                constant: constant_base as u16,
+                arg_derived_e4: arg_base as u16,
+                descriptor: descriptor_base as u16,
+                lookup_additive: lookup_additive_slot.map(|slot| slot as u16),
+            },
+        )?;
+        append_program(&mut desc, &mut next_lane, layer_index, &rewritten)?;
+    }
+
     if uses_table {
         if header.table.is_null() {
             return Err(FwdVmLowerError::MissingTable);
@@ -702,8 +848,408 @@ pub(crate) fn lower_layer_desc(
     }
     desc.mask = mask;
 
-    // ----- geometry. -----
-    desc.count = header.count;
+    Ok(LoweredFwdVm {
+        desc,
+        lookup_additive_slot,
+        decoder_fill_slot,
+    })
+}
 
-    Ok(desc)
+#[cfg(test)]
+mod tests {
+    use std::collections::BTreeMap;
+    use std::path::Path;
+    use std::sync::Arc;
+
+    use gpu_core::primitives::field::{BF, E4};
+    use gpu_gkr_compiler::{
+        ForwardDstLine as DstLine, ForwardInstr as Instr, ForwardLdcSub as LdcSub,
+        ForwardOperandField as OperandField, ForwardOperandLine as OperandLine,
+        ForwardProgram as Program, ForwardSpecialStrategy,
+    };
+    use gpu_trace::witness::circuit_type::{
+        CircuitType, DelegationCircuitType, UnrolledCircuitType, UnrolledMemoryCircuitType,
+        UnrolledNonMemoryCircuitType,
+    };
+
+    use super::*;
+    use crate::programs::GkrPrograms;
+    use crate::upstream::{GKRAddress, GKRCircuitArtifact, ReadPlace};
+
+    fn source_use(
+        layer: usize,
+        compiler_column: u16,
+        matrix_base: usize,
+        matrix_column: usize,
+    ) -> PhysicalSourceUse {
+        PhysicalSourceUse {
+            key: SourceUseKey {
+                layer,
+                window: 0,
+                column: compiler_column,
+            },
+            is_e4: false,
+            matrix_base: matrix_base as *mut u8,
+            stride_bytes: 16,
+            matrix_column,
+        }
+    }
+
+    #[test]
+    fn physical_sources_pack_into_five_hundred_twelve_column_windows() {
+        let uses = [
+            source_use(0, 0, 0x1000, 0),
+            source_use(0, 1, 0x1000, 127),
+            source_use(1, 0, 0x1000, 128),
+            source_use(1, 1, 0x1000, 511),
+            source_use(1, 2, 0x1000, 512),
+        ];
+
+        let geometry = pack_source_geometry(&uses).unwrap();
+        assert_eq!(geometry.n_windows, 2);
+        assert_eq!(geometry.base[0] as usize, 0x1000);
+        assert_eq!(geometry.base[1] as usize, 0x1000 + 512 * 16);
+        assert_eq!(geometry.remap[&uses[0].key], (0, 0));
+        assert_eq!(geometry.remap[&uses[3].key], (0, 511));
+        assert_eq!(geometry.remap[&uses[4].key], (1, 0));
+    }
+
+    #[test]
+    fn seventeenth_physical_source_window_is_rejected() {
+        let uses = (0..17)
+            .map(|index| source_use(index, 0, 0x1000 + index * 0x1000, 0))
+            .collect::<Vec<_>>();
+
+        assert!(matches!(
+            pack_source_geometry(&uses),
+            Err(FwdVmLowerError::SourceWindowOverflow { required: 17 })
+        ));
+    }
+
+    #[test]
+    fn rewrite_rebases_every_layer_local_bank() {
+        let program = Program {
+            instrs: vec![
+                Instr::Add {
+                    field: OperandField::Base,
+                    sign: gpu_gkr_compiler::ForwardSign::Plus,
+                    operands: vec![
+                        OperandLine::Source {
+                            window: 1,
+                            column: 2,
+                        },
+                        OperandLine::Ldc {
+                            sub: LdcSub::Const,
+                            idx: 3,
+                        },
+                        OperandLine::Ldc {
+                            sub: LdcSub::ArgDerivedE4,
+                            idx: 4,
+                        },
+                        OperandLine::Ldc {
+                            sub: LdcSub::ConstDerivedE4,
+                            idx: 0,
+                        },
+                        OperandLine::Special { desc: 5 },
+                        OperandLine::Smem { cell: 6 },
+                        OperandLine::Ldc {
+                            sub: LdcSub::Special,
+                            idx: 1,
+                        },
+                    ],
+                },
+                Instr::Mov {
+                    dir: gpu_gkr_compiler::ForwardMovDir::DstFromAcc,
+                    field: OperandField::Base,
+                    dst: Some(DstLine::GlobalMaterialize { slot: 2, col: 9 }),
+                    src: None,
+                },
+            ],
+        };
+        let mut sources = BTreeMap::new();
+        sources.insert(
+            SourceUseKey {
+                layer: 7,
+                window: 1,
+                column: 2,
+            },
+            (9, 400),
+        );
+        let mut destinations = BTreeMap::new();
+        destinations.insert(
+            DestinationUseKey {
+                layer: 7,
+                slot: 2,
+                column: 9,
+            },
+            (8, 300),
+        );
+        let bases = LayerBankBases {
+            constant: 10,
+            arg_derived_e4: 20,
+            descriptor: 30,
+            lookup_additive: Some(1),
+        };
+
+        let rewritten = rewrite_program(7, &program, &sources, &destinations, bases).unwrap();
+        let Instr::Add { operands, .. } = &rewritten.instrs[0] else {
+            panic!("expected Add")
+        };
+        assert_eq!(
+            operands[0],
+            OperandLine::Source {
+                window: 9,
+                column: 400
+            }
+        );
+        assert_eq!(
+            operands[1],
+            OperandLine::Ldc {
+                sub: LdcSub::Const,
+                idx: 13
+            }
+        );
+        assert_eq!(
+            operands[2],
+            OperandLine::Ldc {
+                sub: LdcSub::ArgDerivedE4,
+                idx: 24,
+            }
+        );
+        assert_eq!(
+            operands[3],
+            OperandLine::Ldc {
+                sub: LdcSub::ConstDerivedE4,
+                idx: 1,
+            }
+        );
+        assert_eq!(operands[4], OperandLine::Special { desc: 35 });
+        assert_eq!(operands[5], OperandLine::Smem { cell: 6 });
+        assert_eq!(
+            operands[6],
+            OperandLine::Ldc {
+                sub: LdcSub::Special,
+                idx: 1,
+            }
+        );
+        assert_eq!(
+            rewritten.instrs[1],
+            Instr::Mov {
+                dir: gpu_gkr_compiler::ForwardMovDir::DstFromAcc,
+                field: OperandField::Base,
+                dst: Some(DstLine::GlobalMaterialize { slot: 8, col: 300 }),
+                src: None,
+            }
+        );
+    }
+
+    #[test]
+    fn encoded_programs_concatenate_and_record_layer_boundaries() {
+        let first = Program {
+            instrs: vec![Instr::Mov {
+                dir: gpu_gkr_compiler::ForwardMovDir::AccFromSrc,
+                field: OperandField::Base,
+                dst: None,
+                src: Some(OperandLine::Smem { cell: 1 }),
+            }],
+        };
+        let second = Program {
+            instrs: vec![
+                Instr::Mov {
+                    dir: gpu_gkr_compiler::ForwardMovDir::AccFromSrc,
+                    field: OperandField::Base,
+                    dst: None,
+                    src: Some(OperandLine::Smem { cell: 2 }),
+                },
+                Instr::Mov {
+                    dir: gpu_gkr_compiler::ForwardMovDir::DstFromAcc,
+                    field: OperandField::Base,
+                    dst: Some(DstLine::Smem { cell: 3 }),
+                    src: None,
+                },
+            ],
+        };
+        let mut desc = empty_desc();
+        let mut next_lane = 0;
+
+        append_program(&mut desc, &mut next_lane, 0, &first).unwrap();
+        append_program(&mut desc, &mut next_lane, 1, &second).unwrap();
+
+        assert_eq!(desc.layers[0].program_offset, 0);
+        assert_eq!(desc.layers[0].instruction_count, 1);
+        assert_eq!(desc.layers[1].program_offset, 2);
+        assert_eq!(desc.layers[1].instruction_count, 2);
+        assert_eq!(next_lane, 6);
+        assert_eq!(&desc.program[..6], &[3, 5, 3, 9, 7, 6]);
+    }
+
+    fn fake_column(place: ReadPlace, field: OperandField) -> ResolvedColumn {
+        let field_index = usize::from(field == OperandField::Ext);
+        let (backing, column) = match place {
+            ReadPlace::BaseLayerMemory { column } => (1, column),
+            ReadPlace::BaseLayerWitness { column } => (2, column),
+            ReadPlace::Setup { column } => (3, column),
+            ReadPlace::Scratch { slot } => (4, slot),
+            ReadPlace::LayerOutput { layer, offset } => (100 + layer * 2 + field_index, offset),
+            ReadPlace::CacheOutput { layer, offset } => (200 + layer * 2 + field_index, offset),
+        };
+        let matrix_base = (0x1000_0000usize + backing * 0x0010_0000) as *mut u8;
+        let stride_bytes = if field == OperandField::Ext { 256 } else { 64 };
+        ResolvedColumn {
+            is_e4: field == OperandField::Ext,
+            ptr: matrix_base.wrapping_add(column * stride_bytes as usize),
+            matrix_base,
+            stride_bytes,
+        }
+    }
+
+    fn lower_embedded_circuit(layout: &str, circuit_type: CircuitType) {
+        let path = Path::new(env!("CARGO_MANIFEST_DIR"))
+            .join("../../cs/compiled_circuits")
+            .join(layout);
+        let artifact: GKRCircuitArtifact<BF> =
+            serde_json::from_slice(&std::fs::read(path).unwrap()).unwrap();
+        let programs = GkrPrograms::compile(circuit_type, Arc::new(artifact)).unwrap();
+
+        let mut resolved = Vec::<(GKRAddress, ResolvedColumn)>::new();
+        let mut insert = |place: ReadPlace, field: OperandField| {
+            let address = read_place_to_gkr_address(&place);
+            let column = fake_column(place, field);
+            if let Some((_, previous)) =
+                resolved.iter().find(|(candidate, _)| *candidate == address)
+            {
+                assert_eq!(previous.is_e4, column.is_e4);
+                assert_eq!(previous.ptr, column.ptr);
+            } else {
+                resolved.push((address, column));
+            }
+        };
+        for layer in &programs.forward.layers {
+            for instruction in &layer.program.instrs {
+                let mut record = |operand: &OperandLine, field: OperandField| {
+                    if let OperandLine::Source { window, column } = *operand {
+                        let place = layer
+                            .source_windows
+                            .resolve_read_place(window, column)
+                            .unwrap();
+                        insert(place, field);
+                    }
+                };
+                match instruction {
+                    Instr::Add {
+                        field, operands, ..
+                    }
+                    | Instr::Mul {
+                        field, operands, ..
+                    } => operands.iter().for_each(|operand| record(operand, *field)),
+                    Instr::Fma {
+                        field_lhs,
+                        field_rhs,
+                        pairs,
+                        ..
+                    } => pairs.iter().for_each(|(lhs, rhs)| {
+                        record(lhs, *field_lhs);
+                        record(rhs, *field_rhs);
+                    }),
+                    Instr::Mov {
+                        field, dst, src, ..
+                    } => {
+                        if let Some(src) = src {
+                            record(src, *field);
+                        }
+                        if let Some(DstLine::GlobalMaterialize { slot, col }) = dst {
+                            let place = layer.backings.slot_col_to_read_place(*slot, *col).unwrap();
+                            insert(place, *field);
+                        }
+                    }
+                }
+            }
+            for special in layer.specials.iter() {
+                if let ForwardSpecialStrategy::PeekDecoder { predicate } = special {
+                    insert(*predicate, OperandField::Base);
+                }
+            }
+        }
+        let resolve = |address: GKRAddress| {
+            resolved
+                .iter()
+                .find_map(|(candidate, column)| (*candidate == address).then_some(*column))
+        };
+        let header = FwdVmInputs {
+            mapping_arena: [1 as *const u32, 2 as *const u32, 3 as *const u32],
+            decoder_mapping_col: Some(0),
+            table: 4 as *const E4,
+            table_len: 16,
+            count: 16,
+            inits_and_teardowns_top_bits: &[0; 32],
+        };
+
+        lower_desc(&programs.forward.layers, &header, &resolve, &|_| unsafe {
+            core::mem::zeroed()
+        })
+        .unwrap();
+    }
+
+    #[test]
+    fn cpu_all_embedded_forward_circuits_fit_one_descriptor() {
+        use DelegationCircuitType::*;
+        use UnrolledCircuitType::*;
+        use UnrolledMemoryCircuitType::*;
+        use UnrolledNonMemoryCircuitType::*;
+
+        let cases = [
+            (
+                "add_sub_lui_auipc_mop_layout_gkr.json",
+                CircuitType::Unrolled(NonMemory(AddSubLuiAuipcMop)),
+            ),
+            (
+                "jump_branch_slt_layout_gkr.json",
+                CircuitType::Unrolled(NonMemory(JumpBranchSlt)),
+            ),
+            (
+                "unsigned_mul_div_layout_gkr.json",
+                CircuitType::Unrolled(NonMemory(MulDivUnsigned)),
+            ),
+            (
+                "shift_binop_layout_gkr.json",
+                CircuitType::Unrolled(NonMemory(ShiftBinary)),
+            ),
+            (
+                "mem_word_only_layout_gkr.json",
+                CircuitType::Unrolled(Memory(LoadStoreWordOnly)),
+            ),
+            (
+                "mem_subword_only_layout_gkr.json",
+                CircuitType::Unrolled(Memory(LoadStoreSubwordOnly)),
+            ),
+            (
+                "inits_and_teardowns_layout_gkr.json",
+                CircuitType::Unrolled(InitsAndTeardowns),
+            ),
+            (
+                "unified_reduced_machine_layout_gkr.json",
+                CircuitType::Unrolled(Unified),
+            ),
+            (
+                "bigint_with_extended_control_layout_gkr.json",
+                CircuitType::Delegation(BigIntWithControl),
+            ),
+            (
+                "keccak_special5_layout_gkr.json",
+                CircuitType::Delegation(KeccakSpecial5),
+            ),
+            (
+                "blake2_with_extended_control_layout_gkr.json",
+                CircuitType::Delegation(Blake2WithCompression),
+            ),
+            (
+                "blake2_g_function_layout_gkr.json",
+                CircuitType::Delegation(Blake2GFunction),
+            ),
+        ];
+
+        for (layout, circuit_type) in cases {
+            lower_embedded_circuit(layout, circuit_type);
+        }
+    }
 }
