+++ gpu/gkr/src/programs.rs
@@ -252,9 +252,6 @@ impl GkrPrograms {
 }
 
 #[cfg(test)]
-mod census;
-
-#[cfg(test)]
 mod tests {
     use super::*;
     use gpu_gkr_compiler::{LeanBoundColumn, LeanBoundWindow, LeanSourceBinding};
