+++ gpu/gkr_compiler/src/forward/encode.rs
@@ -3,21 +3,25 @@
 use super::error::EncodeError;
 use super::isa::*;
 
-fn pack_operand(o: OperandLine, source_layout: SourceLayout) -> Result<u16, EncodeError> {
+fn pack_operand(
+    o: OperandLine,
+    source_window_bits: u32,
+    source_column_bits: u32,
+) -> Result<u16, EncodeError> {
     match o {
         OperandLine::LogicalGlobal { slot, col } => {
             Err(EncodeError::UnboundLogicalSource { slot, col })
         }
         OperandLine::Source { window, column } => {
-            if (window as u32) >= 1 << source_layout.window_bits() {
+            if (window as u32) >= 1 << source_window_bits {
                 return Err(EncodeError::SourceWindowOutOfRange(window));
             }
-            if (column as u32) >= 1 << source_layout.column_bits() {
+            if (column as u32) >= 1 << source_column_bits {
                 return Err(EncodeError::SourceColumnOutOfRange(column));
             }
             Ok(TAG_SOURCE
                 | ((window as u16) << SOURCE_WINDOW_SHIFT)
-                | (column << (SOURCE_WINDOW_SHIFT + source_layout.window_bits())))
+                | (column << (SOURCE_WINDOW_SHIFT + source_window_bits)))
         }
         OperandLine::Smem { cell } => {
             if (cell as u32) >= MAX_CELL {
@@ -81,9 +85,10 @@ fn pack_arith_header(
         | ((sign as u16) << SIGN_SHIFT))
 }
 
-pub fn encode_with_source_layout(
+fn encode_with_source_bits(
     p: &Program,
-    source_layout: SourceLayout,
+    source_window_bits: u32,
+    source_column_bits: u32,
 ) -> Result<Vec<u16>, EncodeError> {
     let mut out = Vec::new();
     for instr in &p.instrs {
@@ -101,7 +106,7 @@ pub fn encode_with_source_layout(
                     *sign,
                 )?);
                 for o in operands {
-                    out.push(pack_operand(*o, source_layout)?);
+                    out.push(pack_operand(*o, source_window_bits, source_column_bits)?);
                 }
             }
             Instr::Mul {
@@ -118,7 +123,7 @@ pub fn encode_with_source_layout(
                     sign,
                 )?);
                 for o in operands {
-                    out.push(pack_operand(*o, source_layout)?);
+                    out.push(pack_operand(*o, source_window_bits, source_column_bits)?);
                 }
             }
             Instr::Fma {
@@ -138,8 +143,8 @@ pub fn encode_with_source_layout(
                     *sign,
                 )?);
                 for (l, r) in pairs {
-                    out.push(pack_operand(*l, source_layout)?);
-                    out.push(pack_operand(*r, source_layout)?);
+                    out.push(pack_operand(*l, source_window_bits, source_column_bits)?);
+                    out.push(pack_operand(*r, source_window_bits, source_column_bits)?);
                 }
             }
             Instr::Mov {
@@ -154,13 +159,19 @@ pub fn encode_with_source_layout(
                         | ((*field as u16) << MOV_FIELD_SHIFT),
                 );
                 match dir {
-                    MovDir::AccFromSrc => {
-                        out.push(pack_operand(src.expect("AccFromSrc src"), source_layout)?)
-                    }
+                    MovDir::AccFromSrc => out.push(pack_operand(
+                        src.expect("AccFromSrc src"),
+                        source_window_bits,
+                        source_column_bits,
+                    )?),
                     MovDir::DstFromAcc => out.push(pack_dst(dst.expect("DstFromAcc dst"))?),
                     MovDir::DstFromSrc => {
                         out.push(pack_dst(dst.expect("DstFromSrc dst"))?);
-                        out.push(pack_operand(src.expect("DstFromSrc src"), source_layout)?);
+                        out.push(pack_operand(
+                            src.expect("DstFromSrc src"),
+                            source_window_bits,
+                            source_column_bits,
+                        )?);
                     }
                 }
             }
@@ -170,7 +181,11 @@ pub fn encode_with_source_layout(
 }
 
 pub fn encode(p: &Program) -> Result<Vec<u16>, EncodeError> {
-    encode_with_source_layout(p, DEFAULT_SOURCE_LAYOUT)
+    encode_with_source_bits(p, SOURCE_WINDOW_BITS, SOURCE_COLUMN_BITS)
+}
+
+pub fn encode_runtime(p: &Program) -> Result<Vec<u16>, EncodeError> {
+    encode_with_source_bits(p, RUNTIME_SOURCE_WINDOW_BITS, RUNTIME_SOURCE_COLUMN_BITS)
 }
 
 #[cfg(test)]
@@ -189,34 +204,22 @@ mod tests {
     }
 
     #[test]
-    fn grouped_source_layout_encodes_its_maximum_coordinate() {
-        let grouped = SourceLayout::new(4, 9).unwrap();
+    fn runtime_source_layout_encodes_its_maximum_coordinate() {
         assert_eq!(
-            encode_with_source_layout(&program_with_source(15, 511), grouped),
+            encode_runtime(&program_with_source(15, 511)),
             Ok(vec![3, 32_764]),
         );
     }
 
     #[test]
-    fn grouped_source_layout_rejects_each_overflow_boundary() {
-        let grouped = SourceLayout::new(4, 9).unwrap();
+    fn runtime_source_layout_rejects_each_overflow_boundary() {
         assert_eq!(
-            encode_with_source_layout(&program_with_source(16, 0), grouped),
+            encode_runtime(&program_with_source(16, 0)),
             Err(EncodeError::SourceWindowOutOfRange(16)),
         );
         assert_eq!(
-            encode_with_source_layout(&program_with_source(0, 512), grouped),
+            encode_runtime(&program_with_source(0, 512)),
             Err(EncodeError::SourceColumnOutOfRange(512)),
-        );
-    }
-
-    #[test]
-    fn source_layout_must_fill_the_existing_thirteen_bit_payload() {
-        assert_eq!(SourceLayout::new(4, 8), None);
-        assert_eq!(SourceLayout::new(0, 13), None);
-        assert_eq!(
-            SourceLayout::new(4, 9),
-            Some(SourceLayout::new(4, 9).unwrap())
         );
     }
 }
