+++ gpu/gkr_compiler/src/forward/isa.rs
@@ -1,7 +1,7 @@
 //! Instruction and operand types for the forward-eval VM.
 //!
-//! These constants define the Rust side of the 16-bit wire format;
-//! `eval_vm_isa.cuh` mirrors them for CUDA.
+//! These constants define the compiler's 16-bit instruction format. Runtime
+//! lowering repacks source coordinates for the CUDA descriptor.
 
 pub(crate) const TYPE_BITS: u32 = 2;
 pub(crate) const SOURCE_WINDOW_SHIFT: u32 = TYPE_BITS;
@@ -11,38 +11,8 @@ pub(crate) const SOURCE_COLUMN_SHIFT: u32 = SOURCE_WINDOW_SHIFT + SOURCE_WINDOW_
 pub(crate) const SOURCE_COLUMN_BITS: u32 = 7;
 pub(crate) const MAX_SOURCE_WINDOWS: u32 = 1 << SOURCE_WINDOW_BITS;
 pub const SOURCE_WINDOW_COLUMNS: u32 = 1 << SOURCE_COLUMN_BITS;
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct SourceLayout {
-    window_bits: u32,
-    column_bits: u32,
-}
-
-impl SourceLayout {
-    pub const fn new(window_bits: u32, column_bits: u32) -> Option<Self> {
-        if window_bits > 0 && column_bits > 0 && window_bits + column_bits == 13 {
-            Some(Self {
-                window_bits,
-                column_bits,
-            })
-        } else {
-            None
-        }
-    }
-
-    pub(crate) const fn window_bits(self) -> u32 {
-        self.window_bits
-    }
-
-    pub(crate) const fn column_bits(self) -> u32 {
-        self.column_bits
-    }
-}
-
-pub const DEFAULT_SOURCE_LAYOUT: SourceLayout = SourceLayout {
-    window_bits: SOURCE_WINDOW_BITS,
-    column_bits: SOURCE_COLUMN_BITS,
-};
+pub(crate) const RUNTIME_SOURCE_WINDOW_BITS: u32 = 4;
+pub(crate) const RUNTIME_SOURCE_COLUMN_BITS: u32 = 9;
 // Compiler-private backing and destination coordinates. These are not source-lane fields.
 pub(crate) const SLOT_BITS: u32 = 4; // ≤16 logical backings/layer
 pub(crate) const COL_BITS: u32 = 10; // ≤1024 logical cols/backing
@@ -214,9 +184,10 @@ mod tests {
     use super::*;
 
     const CUDA_ISA: &str = include_str!("../../../gkr/native/gkr/eval_vm_isa.cuh");
+    const CUDA_FORWARD_DESC: &str = include_str!("../../../gkr/native/gkr/forward/fwd_vm.cuh");
 
-    fn cpp_expr(name: &str) -> &str {
-        CUDA_ISA
+    fn cpp_expr<'a>(source: &'a str, name: &str) -> &'a str {
+        source
             .lines()
             .find_map(|line| {
                 let line = line.trim();
@@ -229,7 +200,7 @@ mod tests {
     }
 
     fn assert_literal(name: &str, value: u32) {
-        assert_eq!(cpp_expr(name), value.to_string(), "{name}");
+        assert_eq!(cpp_expr(CUDA_ISA, name), value.to_string(), "{name}");
     }
 
     #[test]
@@ -239,8 +210,6 @@ mod tests {
             ("FWD_VM_HDR_ARITY_BITS", ARITY_BITS),
             ("FWD_VM_MOV_DIR_BITS", MOV_DIR_BITS),
             ("FWD_VM_OPERAND_TAG_BITS", TYPE_BITS),
-            ("FWD_VM_SOURCE_WINDOW_BITS", SOURCE_WINDOW_BITS),
-            ("FWD_VM_SOURCE_COLUMN_BITS", SOURCE_COLUMN_BITS),
             ("FWD_VM_LDC_SUB_BITS", LDC_SUB_BITS),
             ("FWD_VM_DST_TAG_BITS", DST_TAG_BITS),
             ("FWD_VM_DST_SLOT_BITS", SLOT_BITS),
@@ -275,10 +244,6 @@ mod tests {
                 "FWD_VM_MOV_DIR_SHIFT + FWD_VM_MOV_DIR_BITS",
             ),
             ("FWD_VM_SOURCE_WINDOW_SHIFT", "FWD_VM_OPERAND_TAG_BITS"),
-            (
-                "FWD_VM_SOURCE_COLUMN_SHIFT",
-                "FWD_VM_SOURCE_WINDOW_SHIFT + FWD_VM_SOURCE_WINDOW_BITS",
-            ),
             ("FWD_VM_OPERAND_CELL_SHIFT", "FWD_VM_OPERAND_TAG_BITS"),
             ("FWD_VM_OPERAND_DESC_SHIFT", "FWD_VM_OPERAND_TAG_BITS"),
             ("FWD_VM_LDC_SUB_SHIFT", "FWD_VM_OPERAND_TAG_BITS"),
@@ -293,8 +258,21 @@ mod tests {
                 "FWD_VM_DST_SLOT_SHIFT + FWD_VM_DST_SLOT_BITS",
             ),
         ] {
-            assert_eq!(cpp_expr(name), expression, "{name}");
+            assert_eq!(cpp_expr(CUDA_ISA, name), expression, "{name}");
         }
+
+        assert_eq!(
+            cpp_expr(CUDA_FORWARD_DESC, "FWD_VM_SOURCE_WINDOW_BITS"),
+            RUNTIME_SOURCE_WINDOW_BITS.to_string()
+        );
+        assert_eq!(
+            cpp_expr(CUDA_FORWARD_DESC, "FWD_VM_SOURCE_COLUMN_BITS"),
+            RUNTIME_SOURCE_COLUMN_BITS.to_string()
+        );
+        assert_eq!(
+            cpp_expr(CUDA_FORWARD_DESC, "FWD_VM_SOURCE_COLUMN_SHIFT"),
+            "FWD_VM_SOURCE_WINDOW_SHIFT + FWD_VM_SOURCE_WINDOW_BITS"
+        );
 
         assert_eq!(ARITY_SHIFT, OPCODE_BITS);
         assert_eq!(F0_SHIFT, ARITY_SHIFT + ARITY_BITS);
